requirements.txt
aniso8601==7.0.0
asgiref==3.5.2
certifi==2022.9.14
charset-normalizer==2.1.1
Django==3.2
django-environ==0.9.0
django-filter==22.1
django-graphql-jwt==0.3.4
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
idna==3.4
promise==2.3
psycopg2==2.9.1
psycopg2-binary==2.9.1
PyJWT==2.4.0
pytz==2022.2.1
requests==2.28.1
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
urllib3==1.26.12
